+++ EssentialFeed/EssentialFeed/FeedCache/LocalFeedLoader.swift
@@ -7,20 +7,6 @@
 
 import Foundation
 
-private final class FeedCachePolicy {
-    private static let calendar = Calendar(identifier: .gregorian)
-
-    private static var maxCacheAgeInDays: Int { 7 }
-    
-    private init() { }
-    
-    static func validate(_ timestamp: Date, against date: Date) -> Bool {
-        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else { return false }
-        
-        return date < maxCacheAge
-    }
-}
-
 public class LocalFeedLoader: FeedLoader {
     private let store: FeedStore
     private let currentDate: () -> Date
